requirements.txt
paho-mqtt
pdf2image
pillow